+++ FathomUIKit/Extension/UIColor+Extension.swift
@@ -1,0 +1,15 @@
+//
+//  UIColor+Extension.swift
+//  FathomUIKit
+//
+//  Created by Ethanhuang on 2018/7/30.
+//  Copyright © 2018年 Elaborapp Co., Ltd. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    public static let barTint: UIColor = UIColor.black
+    public static let tint: UIColor = #colorLiteral(red: 1, green: 0.5803921569, blue: 0.1450980392, alpha: 1)
+    public static let background: UIColor = #colorLiteral(red: 0.9803921569, green: 0.9803921569, blue: 0.9803921569, alpha: 1)
+}