+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Knil Changelog
+# Fathom Changelog
 
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Add support for modern AASA format
+- Rename from "Knil" to "Fathom"
 
 ## [1.1.0(9)] - 2019-05-26
 ### Added